requirements.txt
numpy==1.17.4

face-recognition==1.2.3
Flask==1.1.1
Pillow==6.2.1
requests==2.22.0
PyYAML==5.1.2



